added InitFunctions file to make main activity clearer
